requirements.txt
fastapi
aiofiles
uvicorn[standard]
pydantic
PyYAML
openai
google-genai
anthropic
customtkinter
python-multipart
